% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar_dimensiones.R
\name{cargar_dimensiones}
\alias{cargar_dimensiones}
\title{Inserta/actualiza dim_geo, dim_tramo y dim_tiempo}
\usage{
cargar_dimensiones(dt, con, file_name = NULL)
}
\arguments{
\item{dt}{data.table resultante de leer_txt_sii()}

\item{con}{Conexión DBI (PostgreSQL)}

\item{file_name}{Nombre del archivo (se usa sólo para mensajes)}
}
\description{
La función detecta en tiempo de ejecución si existen las
columnas mínimas para cada dimensión.
• Si faltan, muestra \code{message()} y omite la carga.
• En dim_tiempo desecha filas con \code{NA} en \code{Anio}.
}
