% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar_hechos_empresas.R
\name{cargar_hechos_empresas}
\alias{cargar_hechos_empresas}
\title{Inserta o actualiza la tabla de hechos f_empresas (vectorizado)}
\usage{
cargar_hechos_empresas(dt, con)
}
\arguments{
\item{dt}{data.table normalizado con columnas
CodRegion, CodProvincia, CodComuna, TramoVentas,
CodRubro, Anio, NumeroEmpresas, MontoVentas,
Trabajadores, Remuneraciones}

\item{con}{Conexion DBI (PostgreSQL)}
}
\description{
Inserta o actualiza la tabla de hechos f_empresas (vectorizado)
}
