% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar_paquetes.R
\name{cargar_paquetes}
\alias{cargar_paquetes}
\title{Carga paquetes necesarios para el análisis}
\usage{
cargar_paquetes(packages)
}
\arguments{
\item{packages}{carga paquetes que de no estar disponibles son instalados}
}
\description{
Carga paquetes necesarios para el análisis
}
