% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_base_si_no_existe.R
\name{crear_base_si_no_existe}
\alias{crear_base_si_no_existe}
\title{Crear base de datos si no existe (usando .env)}
\usage{
crear_base_si_no_existe(nombre_bd, env_path = NULL)
}
\arguments{
\item{nombre_bd}{Nombre de la base de datos a crear}

\item{env_path}{ruta al archivo .env}
}
\description{
Crear base de datos si no existe (usando .env)
}
