% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_esquema_empresas.R
\name{crear_esquema_empresas}
\alias{crear_esquema_empresas}
\title{Crea y documenta la estructura del Data Warehouse}
\usage{
crear_esquema_empresas(con, schema = "public")
}
\arguments{
\item{con}{Conexión DBI (RPostgres) con permisos DDL}

\item{schema}{Esquema destino; por defecto "public"}
}
\description{
Imprime si cada tabla fue creada o ya existía.
}
