% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_empresas_sii.R
\name{descargar_empresas_sii}
\alias{descargar_empresas_sii}
\title{Descarga EMPRESAS.zip desde el SII y lo descomprime (idempotente)}
\usage{
descargar_empresas_sii(dir = here::here("data-raw/sii"))
}
\arguments{
\item{dir}{Carpeta destino; por defecto "data-raw/sii"}
}
\value{
Vector de rutas a los TXT disponibles
}
\description{
Descarga EMPRESAS.zip desde el SII y lo descomprime (idempotente)
}
