% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establecer_conexion.R
\name{establecer_conexion}
\alias{establecer_conexion}
\title{Establece una conexión a PostgreSQL usando credenciales en .env}
\usage{
establecer_conexion(db_name = Sys.getenv("PGDATABASE"), env_path = NULL)
}
\arguments{
\item{db_name}{Nombre de la base de datos a conectar (opcional)}

\item{env_path}{Ruta al archivo .env (por defecto usa here::here())}
}
\value{
Objeto de conexión DBI
}
\description{
Establece una conexión a PostgreSQL usando credenciales en .env
}
