% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Función para obtener metadatos de cada base de datos}
\usage{
get_metadata(dbname, con)
}
\arguments{
\item{dbname}{nombre de la base de datos}

\item{con}{conexión vía DBI a una base de datos}
}
\description{
Función para obtener metadatos de cada base de datos
}
