% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_txt_sii.R
\name{leer_txt_sii}
\alias{leer_txt_sii}
\title{Lee un TXT/TAB del SII y renombra columnas según un mapeo}
\usage{
leer_txt_sii(file, hdr_map = NULL)
}
\arguments{
\item{file}{Ruta al archivo .txt}

\item{hdr_map}{Data.table opcional con columnas
\code{file}, \code{original}, \code{estandar}
devuelto por \code{\link{detectar_claves_sii}}.
Si se suministra, renombra \code{original} → \code{estandar}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   \code{data.table} con nombres homogéneos y metadatos añadidos.
}\if{html}{\out{</div>}}
}
\description{
Lee un TXT/TAB del SII y renombra columnas según un mapeo
}
