% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listar_bases_usuario.R
\name{listar_bases_usuario}
\alias{listar_bases_usuario}
\title{Lista las bases de datos accesibles (excluye sistema)}
\usage{
listar_bases_usuario()
}
\value{
Vector de nombres de bases de datos visibles para el usuario
}
\description{
Lista las bases de datos accesibles (excluye sistema)
}
