% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiene_anio.R
\name{tiene_anio}
\alias{tiene_anio}
\title{Verifica la presencia de un indicador de año en un \code{data.table} o en el nombre de archivo}
\usage{
tiene_anio(dt, file_name)
}
\arguments{
\item{dt}{\code{data.table} ya leído y con nombres originales de columnas.}

\item{file_name}{\code{character(1)}. Nombre (o ruta) del archivo de origen.}
}
\value{
\code{TRUE} si se encuentra algún indicador de año; \code{FALSE} en caso contrario.
}
\description{
La función se utiliza como filtro previo en rutinas ETL para descartar
archivos que no permiten identificar el año de referencia.
}
\details{
\strong{Criterios de aceptación}
\enumerate{
\item El \code{data.table} contiene una columna llamada \strong{\code{Anio}}.
\item El \code{data.table} contiene una columna \strong{\code{AnioComercial}} con al menos
un valor no nulo.
\item El nombre del archivo incluye un patrón \verb{20xx}.
}
}
\examples{
\dontrun{
dt <- data.table::data.table(AnioComercial = 2024, x = 1:10)
tiene_anio(dt, "PUB_ACT_2024.txt")   # devuelve TRUE

dt2 <- data.table::data.table(x = 1:5)
tiene_anio(dt2, "PUB_TRAM_PROV.txt")  # devuelve FALSE
}

}
\keyword{internal}
